chore(build_and_deploy.yml): update GitHub Actions workflow to use the latest versions of dependencies and improve readability chore(build.gradle): add 'application' plugin to build.gradle to enable running the application as a standalone Java application chore(gradle-wrapper.properties): update Gradle distribution URL to use version 8.3 chore(gradlew): update gradlew script to include license information and improve compatibility with different POSIX shells chore(gradlew.bat): update gradlew.bat script to improve compatibility with different Windows shells
The changes in the build_and_deploy.yml file update the GitHub Actions workflow to use the latest versions of dependencies and improve the readability of the script. The build.gradle file adds the 'application' plugin to enable running the application as a standalone Java application. The gradle-wrapper.properties file updates the Gradle distribution URL to use version 8.3. The gradlew and gradlew.bat scripts are updated to include license information and improve compatibility with different POSIX shells and Windows shells, respectively. These changes aim to enhance the build and deployment process and ensure compatibility with the latest tools and platforms.