feat: add volume mapping for log files in Docker container and update logback.xml configuration
The build_docker_and_start.bat script now includes a volume mapping for the log files in the Docker container. This allows the logs to be persisted outside of the container in the specified directory "/var/logs/tekken-offline-tg-bot".

The logback.xml configuration file has been updated to include a new appender named "FILE" which writes log messages to a file in the specified directory. The log file name is set to "log" and it will be rolled over daily with a maximum of 30 days' worth of history. The total size of the log files is capped at 3GB.

These changes improve the logging capabilities of the application by ensuring that log files are persisted and can be accessed outside of the Docker container.